requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.11
aiosignal==1.4.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
asyncpg==0.30.0
attrs==25.4.0
beautifulsoup4==4.12.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
coverage==7.13.1
curl_cffi==0.13.0
dataclasses-json==0.6.7
distro==1.9.0
feedparser==6.0.11
filetype==1.2.0
frozendict==2.4.7
frozenlist==1.8.0
google-ai-generativelanguage==0.6.10
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.3
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.63
langchain-google-genai==2.0.8
langchain-text-splitters==0.3.8
langgraph==0.2.45
langgraph-checkpoint==2.1.2
langgraph-sdk==0.1.74
langsmith==0.2.11
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
openai==1.58.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
pandas==2.3.2
pandas-ta==0.4.71b0
peewee==3.19.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pyparsing==3.3.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.0
pytz==2024.2
PyYAML==6.0.3
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
scipy==1.14.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
webencodings==0.5.1
websockets==16.0
win32_setctime==1.2.0
yarl==1.22.0
yfinance==1.0
